requirements.txt
torch 
torchvision 
torchaudio
transformers
peft 
tqdm 
Levenshtein
pandas 
numpy
einops 
timm
Pillow 
supervision
datasets